Re-add a way to run embed_test_data.go for now
Android have a lot of build systems and we have not yet managed to
wrangle all of them to just consume the test data directly. (Apparently
passing test data to tests is a great unsolved problem in computer
science.)

Since this is mostly a backup mechanism now, and we don't want to apply
this to *all* Android's builds, this flag works differently from before.
Now it just drops crypto_test_data.cc somewhere. It is the caller's
responsibility to link this into crypto_test and also build test_support
with -DBORINGSSL_CUSTOM_GET_TEST_DATA.

Bug: 542
Change-Id: I7faad4fa80dfacfdadd533b3cd7de95bf9e258d4