More ECH certificates
In the rejection tests, make sure the certificate served matches the
public name in the ECH config, so shims which verify the returned
certificate can make sure they are using the right name.

Previously there was a mismatch where the name to be verified was
public.example, but the certificate returned was valid for "test".

Change-Id: I511415431e01c9a83766c633cf11c34f0fa93058