Adds functionality for instantiating and using `Algorithm`
This CL moves some methods that exist in every hashing algorithm implementation
to the `Algorithm` trait; namely, `new`, `update`, and `digest`. This allows for
clients to instantiate the hashing algorithms, and to call `update` and `digest`
on objects that have this trait. It allows a generic representation of an algorithm
that can be used.

This CL also adds the block size as a const field in the `Algorithm` trait.

Note: this specifically supports the hashing algorithm use in the style of what is
done in AOSP's packages/modules/Virtualization/libs/apkverify, which was previously
using the rust openssl `Hasher` to instantiate algorithms specified by their `MessageDigest`.

This CL is a WIP as we iterate on the design.

Change-Id: Ic47691ee2a4303923519b246de7d9724da90f60d