Revert "Adds functionality for instantiating and using `Algorithm`" and friends.
This reverts the following commits:
e1a860c3745c77cb83228dde1b73fa62eaf43930
a11277e187e407d0ef403b8a60d9a32eaab7d301
25cf1bb965ba9ae0302cbc6de4ff4dd6cdbbc016

There were a couple nuisances caused by this change, stemming from a
Rust language deficiency. If you move a method from a type to a trait,
this is a backwards-incompatible change that is visible to the caller.
That is, Rust does not allow introducing abstractions in a
backwars-compatible way!

This meant you had to import Algorithm to use a hash. This made updating
downstream code hard, but at least par for the course with Rust. It
seems Rust just isn't a language where API stability and library
evolution are possible.

However, even after we pay the transition costs, downstream code needed
to import WithOutputLength. This is too much of an implementation detail
to leak into the public API, so revert it for now. We'll need to find
some way to do this that keeps the public API reasonable.

Change-Id: I82d00b47a77fe77b5893b1e9b15faef727ef9866