Rewrite CBS_get_asn1_int64 slightly
GCC 13.2.0 has a false positive in -Wstringop-overflow. Oddly, I can
only reproduce it with -O2 -march=native. The old code was also correct,
but this version seems to do a better job of avoiding the warning.
Instead of reversing the variable-length string while sign-exnteding, we
assemble a big-endian sign-extended version and then do a fixed-width
byte swap at the end.

Fixed: 42290598
Change-Id: I6d5de1e1d6d117f6b5947d3a2155e794764eb472