Split off DTLS record header parsing in test runner.
The DTLS 1.3 record header is formatted differently than the old record
header, but the code to read/process a DTLS record mixes record header
parsing with other record processing code. This change provides a clear
delineation between processing the record header and processing the
record, which will assist in adding support for the DTLS 1.3 record
header.

Bug: 715
Change-Id: I13a0bb5c184e79b88f064e9ac8ecbc82eb56750a