Add virtual destructors.
b7f5443cfc added an abstract class without a virtual destructor, which
is fine in that context, but some places build with -Wnon-virtual-dtor
and thus:

  error: 'bssl::AesHwCipherScorer' has virtual functions but non-virtual destructor

Thus reflect this build configuration in our CMake build and add the
needed lines to satisfy the compiler.

Also put the classes in an anonymous namespace to more strongly hint, if
that is needed, that it can all be optimised down to almost nothing,
which it hopefully is.

Change-Id: Icd51782641073f55d1672dcf6a6adb82e3b27fb4