Remove resp variable in BN_div
It's much clearer if we just reference res->d[i] directly. Note that the
removed res->neg clearing is a no-op because bn_set_minimal_width fills
the value in anyway. It was also impossible for res->width to be zero
because of the resizing step (see the bn_resize_words call). Even if it
were possible for it to be zero, that would mean the loop doesn't run,
and the resp pointer was only read outside the loop. So we can treat the
function as if it unconditionally decremented resp.

Bug: 358687140
Change-Id: I5e2d4ca03fd808cacd4f4647843a7894bf7a2f05