Clean up the tail of the BN_div loop
I'm not sure what the comment about overflowing when q = 0 is about. The
bounds in "the first part of the loop" imply that we've either computed
q+1 or q and the borrow check exactly captures the q+1 case.

Moreover, this addition is expected to *always* overflow. It cancels out
the underflow from subtracting too many.

Bug: 358687140
Change-Id: I24bf8c9c37dcd1145667d7f0e8457c0e63e8783c